Add files via upload
This is a fun little project i worked on while I was less busy involving the prediction of adult mortality using a simple linear regression. I applied the use of Mean Squared Error and R-squared score metrics to quantify how well the regression model fit the data and to indicate the proportion  of the variance that is predictable from the independent variables.